backend/requirements copy.txt
supabase==1.0.3
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.1
pinecone-client==2.2.4
langchain==0.0.335
langchain_pinecone==0.1.1
langchain-community==0.0.17
streamlit==1.28.0
agno==0.1.1
pydantic==2.4.2
bs4==0.0.1
requests==2.31.0
aiofiles==23.2.1
streamlit